+++ package.json
@@ -12,10 +12,12 @@
   "dependencies": {
     "@originprotocol/origin-storybook": "^0.27.2",
     "@types/he": "^1.1.2",
+    "@types/numeral": "^2.0.2",
     "@types/qs": "^6.9.7",
     "@types/sanitize-html": "^2.8.0",
     "critters": "^0.0.16",
     "dotenv": "^16.0.2",
+    "ethers": "^6.0.3",
     "he": "^1.2.0",
     "moment": "^2.29.4",
     "next": "12.3.0",
